refactor(base): move specific classes to application project
